pyproject.toml
[tool.poetry]
name = "linx-bot"
version = "0.1.0"
description = ""
authors = ["l4blee <76159455+l4blee@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = {extras = ["srv"], version = "^4.0.1"}
vk-api = "^11.9.7"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
